+++ src/Inline/Parser/AutolinkParser.php
@@ -18,7 +18,7 @@ use League\CommonMark\Inline\Element\Link;
 use League\CommonMark\InlineParserContext;
 use League\CommonMark\Util\UrlEncoder;
 
-class AutolinkParser implements InlineParserInterface
+final class AutolinkParser implements InlineParserInterface
 {
     const EMAIL_REGEX = '/^<([a-zA-Z0-9.!#$%&\'*+\\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*)>/';
     const OTHER_LINK_REGEX = '/^<[A-Za-z][A-Za-z0-9.+-]{1,31}:[^<>\x00-\x20]*>/i';