+++ src/Inline/Renderer/LinkRenderer.php
@@ -22,7 +22,7 @@ use League\CommonMark\Util\ConfigurationAwareInterface;
 use League\CommonMark\Util\ConfigurationInterface;
 use League\CommonMark\Util\RegexHelper;
 
-class LinkRenderer implements InlineRendererInterface, ConfigurationAwareInterface
+final class LinkRenderer implements InlineRendererInterface, ConfigurationAwareInterface
 {
     /**
      * @var ConfigurationInterface